+++ crates/core/task/src/lib.rs
@@ -2,6 +2,7 @@ mod errors;
 mod file_group;
 mod target;
 mod task;
+mod task_options;
 pub mod test;
 mod token;
 mod types;
@@ -10,6 +11,7 @@ pub use errors::*;
 pub use file_group::FileGroup;
 pub use moon_config::{PlatformType, TargetID, TaskConfig, TaskID, TaskOptionsConfig};
 pub use target::{Target, TargetProjectScope};
-pub use task::{Task, TaskOptions};
+pub use task::*;
+pub use task_options::*;
 pub use token::{ResolverData, ResolverType, TokenResolver, TokenType};
 pub use types::*;
